+++ Forge/src/main/java/com/christofmeg/brutalharvest/common/block/LettuceCropBlock.java
@@ -24,30 +24,24 @@ import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.StateDefinition;
 import net.minecraft.world.level.block.state.properties.IntegerProperty;
 import net.minecraft.world.level.gameevent.GameEvent;
-import net.minecraft.world.level.storage.loot.LootParams;
 import net.minecraft.world.phys.BlockHitResult;
 import net.minecraft.world.phys.shapes.CollisionContext;
 import net.minecraft.world.phys.shapes.VoxelShape;
 import org.jetbrains.annotations.NotNull;
 
-import java.util.List;
+public class LettuceCropBlock extends CropBlock {
 
-public class TomatoCropBlock extends CropBlock {
-
-    public static final IntegerProperty AGE = IntegerProperty.create("age", 0, 8);
+    public static final IntegerProperty AGE = IntegerProperty.create("age", 0, 5);
     private static final VoxelShape[] SHAPE_BY_AGE = new VoxelShape[] {
+            Block.box(0.0, 0.0, 0.0, 16.0, 3.0, 16.0),
             Block.box(0.0, 0.0, 0.0, 16.0, 4.0, 16.0),
-            Block.box(0.0, 0.0, 0.0, 16.0, 8.0, 16.0),
-            Block.box(0.0, 0.0, 0.0, 16.0, 11.0, 16.0),
-            Block.box(0.0, 0.0, 0.0, 16.0, 13.0, 16.0),
-            Block.box(0.0, 0.0, 0.0, 16.0, 14.0, 16.0),
-            Block.box(0.0, 0.0, 0.0, 16.0, 14.0, 16.0),
-            Block.box(0.0, 0.0, 0.0, 16.0, 14.0, 16.0),
-            Block.box(0.0, 0.0, 0.0, 16.0, 14.0, 16.0),
-            Block.box(0.0, 0.0, 0.0, 16.0, 14.0, 16.0)
+            Block.box(0.0, 0.0, 0.0, 16.0, 6.0, 16.0),
+            Block.box(0.0, 0.0, 0.0, 16.0, 9.0, 16.0),
+            Block.box(0.0, 0.0, 0.0, 16.0, 12.0, 16.0),
+            Block.box(0.0, 0.0, 0.0, 16.0, 12.0, 16.0)
     };
 
-    public TomatoCropBlock(Properties properties) {
+    public LettuceCropBlock(Properties properties) {
         super(properties);
         this.registerDefaultState(this.stateDefinition.any().setValue(this.getAgeProperty(), 0));
     }
@@ -58,19 +52,19 @@ public class TomatoCropBlock extends CropBlock {
     }
 
     @Override
-    public void randomTick(@NotNull BlockState state, @NotNull ServerLevel $$1, @NotNull BlockPos $$2, @NotNull RandomSource $$3) {
+    public void randomTick(@NotNull BlockState state, @NotNull ServerLevel level, @NotNull BlockPos pos, @NotNull RandomSource random) {
         int age = state.getValue(AGE);
-        if (age < 8 && $$3.nextInt(5) == 0 && $$1.getRawBrightness($$2.above(), 0) >= 9) {
-            BlockState $$5 = state;
-            if (age == 7) {
-                if ($$3.nextInt(4) == 0) { // 25% chance
-                    $$5 = state.setValue(AGE, age + 1);
+        if (age < this.getMaxAge() + 1 && random.nextInt(5) == 0 && level.getRawBrightness(pos.above(), 0) >= 9) {
+            if (age == this.getMaxAge()) {
+                if (random.nextInt(4) == 0) { // 25% chance
+                    state = state.setValue(AGE, age + 1);
                 }
             } else {
-                $$5 = state.setValue(AGE, age + 1);
+                state = state.setValue(AGE, age + 1);
             }
-            $$1.setBlock($$2, $$5, 2);
-            $$1.gameEvent(GameEvent.BLOCK_CHANGE, $$2, GameEvent.Context.of($$5));
+            state = state.setValue(AGE, age + 1);
+            level.setBlock(pos, state, 2);
+            level.gameEvent(GameEvent.BLOCK_CHANGE, pos, GameEvent.Context.of(state));
         }
     }
 
@@ -92,27 +86,12 @@ public class TomatoCropBlock extends CropBlock {
 
     @Override
     public int getMaxAge() {
-        return 7;
-    }
-
-    @SuppressWarnings("deprecation")
-    @Override
-    public @NotNull List<ItemStack> getDrops(@NotNull BlockState state, LootParams.@NotNull Builder builder) {
-        List<ItemStack> dropsOriginal = super.getDrops(state, builder);
-        int randomTomatoes = 3 + builder.getLevel().random.nextInt(4);
-        if (!dropsOriginal.isEmpty()) {
-            return dropsOriginal;
-        }
-        return
-                state.getValue(AGE) == 4 || state.getValue(AGE) == 5 || state.getValue(AGE) == 6 ?  List.of(new ItemStack(ItemRegistry.UNRIPE_TOMATO.get(), randomTomatoes), new ItemStack(this, 2)) :
-                        state.getValue(AGE) == 7 ? List.of(new ItemStack(ItemRegistry.TOMATO.get(), randomTomatoes), new ItemStack(this, 2)) :
-                                state.getValue(AGE) == 8 ? List.of(new ItemStack(ItemRegistry.ROTTEN_TOMATO.get(), randomTomatoes), new ItemStack(this, 2)) :
-                                        List.of(new ItemStack(this, 1));
+        return 4;
     }
 
     @Override
     protected @NotNull ItemLike getBaseSeedId() {
-        return ItemRegistry.TOMATO_SEEDS.get();
+        return ItemRegistry.LETTUCE_SEEDS.get();
     }
 
     @Override
@@ -128,24 +107,24 @@ public class TomatoCropBlock extends CropBlock {
 
     @Override
     public boolean isValidBonemealTarget(@NotNull LevelReader levelReader, @NotNull BlockPos pos, BlockState state, boolean $$3) {
-        return state.getValue(AGE) < 6;
+        return state.getValue(AGE) < this.getMaxAge();
     }
 
     @SuppressWarnings("deprecation")
     @Override
     public @NotNull InteractionResult use(BlockState state, @NotNull Level level, @NotNull BlockPos pos, @NotNull Player player, @NotNull InteractionHand interactionHand, @NotNull BlockHitResult blockHitResult) {
         int age = state.getValue(AGE);
-        boolean reachedTomatoAge = age == 4;
-        if (!reachedTomatoAge && player.getItemInHand(interactionHand).is(Items.BONE_MEAL)) {
+        boolean reachedHarvestAge = age == this.getMaxAge();
+        if (!reachedHarvestAge && player.getItemInHand(interactionHand).is(Items.BONE_MEAL)) {
             return InteractionResult.PASS;
         } else if (age > 3) {
             ItemStack stack = player.getItemInHand(interactionHand);
             if (stack.getItem() instanceof KnifeItem) {
-                int randomTomatoes = 3 + level.random.nextInt(4);
+                int randomSeeds = level.random.nextInt(2);
                 popResource(level, pos,
-                        age == 4 || age == 5 || age == 6 ? new ItemStack(ItemRegistry.UNRIPE_TOMATO.get(), randomTomatoes) :
-                                age == 7 ? new ItemStack(ItemRegistry.TOMATO.get(), randomTomatoes) :
-                                        age == 8 ? new ItemStack(ItemRegistry.ROTTEN_TOMATO.get(), randomTomatoes) : ItemStack.EMPTY);
+                        age == 4 ? new ItemStack(ItemRegistry.LETTUCE.get()) :
+                                age == 5 ? new ItemStack(this, randomSeeds) :
+                                        ItemStack.EMPTY);
                 level.playSound(null, pos, SoundEvents.ITEM_PICKUP, SoundSource.BLOCKS, 1.0F, 0.8F + level.random.nextFloat() * 0.4F);
                 BlockState newBlockState = state.setValue(AGE, 3);
                 level.setBlock(pos, newBlockState, 2);