+++ Forge/changelog.md
@@ -1,4 +1,4 @@
-1.6.1:
+1.7.0:
 * Added on-hit mob effects to rotten tomatoes
 * Added Strawberry food item
 * Added Blueberry food item
@@ -9,3 +9,7 @@
 * Added jar, honey jar, strawberry jam and blueberry jam
 * Added toast with Fried egg, scrambled egg and boiled egg
 * Added toast with honey jar, strawberry jam and blueberry jam
+* Fix Tomato Crop not random ticking to rotten age
+* Added Lettuce Crop
+* Added craft remainders to Corn Kernel Bowl, Honey Jar, Blueberry Jam, Strawberry Jam, Rubber Bucket and Popcorn
+* Updated Corn item texture